pyproject.toml
[project]
name = "ocr-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "opencv-python>=4.12.0.88",
    "opencv-contrib-python>=4.12.0.88",
    "requests>=2.32.5",
    "pytesseract>=0.3.10",
    "pdf2image>=1.17.0",
    "pillow>=10.0.0",
    "openai>=1.40.0",
    "python-dotenv>=1.0.1",
    "openpyxl>=3.1.5",
    "tqdm>=4.66.4",
]
